Normative: Consistently call observable operations with undefined options
Where possible, observable calls originating from within Temporal, that
require an options argument, should pass `undefined` as that options
argument, rather than `{}` or `Object.create(null)`.

This allows the options parameter to be omitted when calling the abstract
operations CalendarDateAdd, DateFromFields, MonthDayFromFields, and
YearMonthFromFields, and assumes that it is `undefined` if omitted, so
that user code sees `undefined` as well when the options value is passed
along.

Whenever one of these abstract operations is invoked from within Temporal
with the default options, pass `undefined` instead of
! OrdinaryObjectCreate(null).

This also requires changing ToTemporalOverflow, ToTemporalDisambiguation,
and ToTemporalOffset to accept `undefined` in addition to a normalized
options object.

Closes: #1685.